# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Creative Studio CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Archive user (soft-delete, admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Archived user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /services:
    get:
      operationId: listServices
      tags: [analytics]
      summary: List all services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [analytics]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceBody"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    delete:
      operationId: deleteService
      tags: [analytics]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /workflow-templates:
    get:
      operationId: listWorkflowTemplates
      tags: [workflow-templates]
      summary: List all workflow templates with their milestones
      responses:
        "200":
          description: List of workflow templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowTemplate"
    post:
      operationId: createWorkflowTemplate
      tags: [workflow-templates]
      summary: Create a new workflow template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkflowTemplateBody"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowTemplate"
  /workflow-templates/{id}:
    patch:
      operationId: updateWorkflowTemplate
      tags: [workflow-templates]
      summary: Update a workflow template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkflowTemplateBody"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowTemplate"
    delete:
      operationId: deleteWorkflowTemplate
      tags: [workflow-templates]
      summary: Delete a workflow template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /workflow-templates/{id}/milestones:
    post:
      operationId: addTemplateMilestone
      tags: [workflow-templates]
      summary: Add a milestone to a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateMilestoneBody"
      responses:
        "201":
          description: Created milestone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateMilestone"
  /workflow-templates/{id}/milestones/{milestoneId}:
    patch:
      operationId: updateTemplateMilestone
      tags: [workflow-templates]
      summary: Update a template milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: milestoneId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateMilestoneBody"
      responses:
        "200":
          description: Updated milestone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateMilestone"
    delete:
      operationId: deleteTemplateMilestone
      tags: [workflow-templates]
      summary: Delete a template milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: milestoneId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, photographer, client]
        profession:
          type: ["string", "null"]
        paymentType:
          type: string
          enum: [monthly_salary, per_project]
        salaryAmount:
          type: ["number", "null"]
        canViewFinancials:
          type: boolean
        canManageClients:
          type: boolean
        canManageAllProjects:
          type: boolean
        canInvoice:
          type: boolean
        canViewLeads:
          type: boolean
        canViewAccounting:
          type: boolean
        archivedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - username
        - name
        - role
        - paymentType
        - canViewFinancials
        - canManageClients
        - canManageAllProjects
        - canInvoice
        - canViewLeads
        - canViewAccounting
        - createdAt
    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, photographer, client]
        profession:
          type: ["string", "null"]
        paymentType:
          type: string
          enum: [monthly_salary, per_project]
        salaryAmount:
          type: ["number", "null"]
        canViewFinancials:
          type: boolean
        canManageClients:
          type: boolean
        canManageAllProjects:
          type: boolean
        canInvoice:
          type: boolean
        canViewLeads:
          type: boolean
        canViewAccounting:
          type: boolean
      required:
        - username
        - password
        - name
        - role
    UpdateUserBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        password:
          type: ["string", "null"]
        profession:
          type: ["string", "null"]
        paymentType:
          type: ["string", "null"]
        salaryAmount:
          type: ["number", "null"]
        canViewFinancials:
          type: ["boolean", "null"]
        canManageClients:
          type: ["boolean", "null"]
        canManageAllProjects:
          type: ["boolean", "null"]
        canInvoice:
          type: ["boolean", "null"]
        canViewLeads:
          type: ["boolean", "null"]
        canViewAccounting:
          type: ["boolean", "null"]
...
    Service:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        createdAt:
          type: string
      required:
        - id
        - title
        - price
        - createdAt
    CreateServiceBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
      required:
        - title
        - price
    TemplateMilestone:
      type: object
      properties:
        id:
          type: integer
        templateId:
          type: integer
        title:
          type: string
        titleAr:
          type: ["string", "null"]
        titleFr:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        order:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - templateId
        - title
        - order
        - createdAt
    WorkflowTemplate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/TemplateMilestone"
        createdAt:
          type: string
      required:
        - id
        - name
        - milestones
        - createdAt
    CreateWorkflowTemplateBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        milestones:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
              titleAr:
                type: ["string", "null"]
              titleFr:
                type: ["string", "null"]
              description:
                type: ["string", "null"]
              order:
                type: integer
            required:
              - title
      required:
        - name
    UpdateWorkflowTemplateBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
    CreateTemplateMilestoneBody:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: ["string", "null"]
        titleFr:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        order:
          type: integer
      required:
        - title
...